+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Add `configuration_sources` option to config classes to specify which data sources to use. ([@palkan][])
+
 ## 2.7.2 (2025-04-01)
 
 - Address NameError: instance variable @secrets not defined issue ([@fbuys][])